Make ITransactionControl use the IManagedStoreConnection directly. That way we take advantage of the superior rollback handling.
