+++ src/relstorage/adapters/tests/test_txncontrol.py
@@ -16,11 +16,24 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
+from hamcrest import assert_that
+from nti.testing.matchers import validly_provides
+
 from relstorage.tests import TestCase
 from relstorage.tests import MockConnection
 from relstorage.tests import MockConnectionManager
 from relstorage.tests import MockCursor
 from relstorage.tests import MockPoller
+
+from ..interfaces import ITransactionControl
+
+class MockStoreConnection(object):
+
+    def __init__(self, conn):
+        self.connection = conn
+
+    def rollback_quietly(self):
+        self.connection.rollback()
 
 class TestTransactionControl(TestCase):
 
@@ -36,6 +49,9 @@ class TestTransactionControl(TestCase):
     def _makeOne(self, keep_history=True, binary=None):
         return self._getClass()(MockConnectionManager(), MockPoller(),
                                 keep_history, binary or self.Binary)
+
+    def test_provides(self):
+        assert_that(self._makeOne(), validly_provides(ITransactionControl))
 
     def test_get_tid_empty_db(self):
         inst = self._makeOne()
@@ -65,17 +81,17 @@ class TestTransactionControl(TestCase):
 
     def test_commit_phase1(self):
         inst = self._makeOne()
-        result = inst.commit_phase1(None, None, None)
+        result = inst.commit_phase1(None, None)
         self.assertEqual(result, '-')
 
     def test_commit_phase2(self):
         inst = self._makeOne()
         conn = MockConnection()
-        inst.commit_phase2(conn, None, None)
+        inst.commit_phase2(MockStoreConnection(conn), None)
         self.assertTrue(conn.committed)
 
     def test_abort(self):
         inst = self._makeOne()
         conn = MockConnection()
-        inst.abort(conn, None, None)
+        inst.abort(MockStoreConnection(conn), None)
         self.assertTrue(conn.rolled_back)
