+++ src/relstorage/storage/tpc/finish.py
@@ -32,8 +32,7 @@ def Finish(vote_state):
     txn = vote_state.prepared_txn
     assert txn is not None
     vote_state.adapter.txncontrol.commit_phase2(
-        vote_state.store_connection.connection,
-        vote_state.store_connection.cursor,
+        vote_state.store_connection,
         txn)
     vote_state.committing_tid_lock.release_commit_lock(vote_state.store_connection.cursor)
     vote_state.cache.after_tpc_finish(vote_state.committing_tid_lock.tid)
